added implementation of 'Watch' for collection + tested all platforms works
