+++ lib/database/mongo_document.dart
@@ -3,6 +3,9 @@ import 'dart:collection';
 import 'dart:convert';
 
 import 'package:bson/bson.dart';
+import 'package:flutter/foundation.dart';
+
+import '../plugin_support.dart';
 
 /// A representation of a document as a Map
 class MongoDocument {
@@ -67,9 +70,14 @@ class MongoDocument {
   }
 
   /// Parses a string in MongoDB Extended JSON format to a Document
-  static MongoDocument parse(String jsonString) {
-    Map<String, dynamic> map = json.decode(jsonString);
-
+  static MongoDocument parse(data) {
+    Map<String, dynamic> map = {};
+    if(kIsWeb){
+      map = FlutterMongoStitch.customEventToMap(data);
+    }
+    else {
+      map = json.decode(data);
+    }
     // fix MongoDB bullshit
     map.forEach((key, value) {
       if (value is LinkedHashMap) {
