+++ lib/flutter_mongo_stitch.dart
@@ -1,4 +1,4 @@
-export 'plugin.dart' hide FlutterMongoStitch;
+export "plugin.dart" hide FlutterMongoStitch;
 export 'auth/auth.dart';
 export 'database/database.dart';
 export 'mongo_stitch_client.dart';
