+++ lib/plugin_support.dart
@@ -1,0 +1,3 @@
+export 'plugin_mobile_support.dart'
+  if (dart.library.html) 'plugin_web_support.dart'
+  if (dart.library.io) 'plugin_mobile_support.dart';
